文件读取采用流输入。另外，segment的double类型换成float类型。

+++ sense-voice/csrc/common.h
@@ -300,7 +300,7 @@ struct sense_voice_segment {
     size_t t1;                   // 时间区间右端点
     // std::string text;         // tokens对应的文本
     std::vector<int> tokens;     // 识别后的tokens
-    std::vector<double> samples;  // 具体音频
+    std::vector<float> samples;  // 具体音频
     // std::vector<float> 
     // bool speaker_turn_next;
 };
